+++ src/SelectScreen.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import { View, Text, Button} from 'react-native';
+
+export default class SelectScreen extends Component {
+  render() {
+    return (
+      <View>
+        <Text style={{fontSize:30}}>여기에 체크박스랑 task.text 데려와야해요</Text>
+      </View>
+    );
+  }
+}
